+++ app/api/log.js
@@ -1,0 +1,25 @@
+const express = require('express');
+const nocache = require('nocache');
+const { getLogLevel } = require('../configuration');
+
+const router = express.Router();
+
+function getLog(req, res) {
+    res.status(200).json({
+        level: getLogLevel(),
+    });
+}
+
+/**
+ * Init Router.
+ * @returns {*}
+ */
+function init() {
+    router.use(nocache());
+    router.get('/', getLog);
+    return router;
+}
+
+module.exports = {
+    init,
+};