+++ app/api/trigger.js
@@ -1,0 +1,14 @@
+const { triggers } = require('../registry');
+const component = require('./component');
+
+/**
+ * Init Router.
+ * @returns {*}
+ */
+function init() {
+    return component.init(triggers);
+}
+
+module.exports = {
+    init,
+};