+++ app/configuration/index.js
@@ -1,5 +1,8 @@
 const envProp = require('env-dot-prop');
 
+function getLogLevel() {
+    return process.env.WUD_LOG_LEVEL || 'info';
+}
 /**
  * Get Input configurations.
  */
@@ -21,8 +24,17 @@ function getStoreConfiguration() {
     return envProp.get('wud.store');
 }
 
+/**
+ * Get API configurations.
+ */
+function getApiConfiguration() {
+    return envProp.get('wud.api');
+}
+
 module.exports = {
+    getLogLevel,
     getStoreConfiguration,
     getWatcherConfigurations,
     getTriggerConfigurations,
+    getApiConfiguration,
 };