+++ e2e/config/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    protocol: process.env.PROTOCOL || 'http',
+    host: process.env.HOST || 'localhost',
+    port: process.env.PORT || 3000,
+};