+++ e2e/features/support/init.js
@@ -1,0 +1,7 @@
+const apickli = require('apickli');
+const { Before } = require('cucumber');
+const configuration = require('../../config');
+
+Before(function initApickli() {
+    this.apickli = new apickli.Apickli(configuration.protocol, `${configuration.host}:${configuration.port}`);
+});